Another namespaces fix commit.
