+++ src/CoiniumServ/Coin/Configs/CoinConfigFactory.cs
@@ -16,7 +16,7 @@
  *   along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-using Coinium.Common.Config;
+using Coinium.Common.Configuration;
 using Coinium.Common.Context;
 
 namespace Coinium.Coin.Configs
