+++ src/CoiniumServ/Coin/Configs/ICoinConfig.cs
@@ -16,7 +16,7 @@
  *   along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-using Coinium.Common.Config;
+using Coinium.Common.Configuration;
 
 namespace Coinium.Coin.Configs
 {
