+++ src/CoiniumServ/Coin/Daemon/Config/IDaemonConfig.cs
@@ -17,7 +17,7 @@
 */
 
 using System;
-using Coinium.Common.Config;
+using Coinium.Common.Configuration;
 
 namespace Coinium.Coin.Daemon.Config
 {
