+++ src/CoiniumServ/CoiniumServ.csproj
@@ -85,7 +85,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Common\Context\ApplicationContext.cs" />
-    <Compile Include="Common\Config\JsonConfigReader.cs" />
+    <Compile Include="Common\Configuration\JsonConfigReader.cs" />
     <Compile Include="Rpc\Service\RpcServiceNames.cs" />
     <Compile Include="Common\Extensions\ArrayExtensions.cs" />
     <Compile Include="Common\Extensions\EnumerableExtensions.cs" />
@@ -119,7 +119,7 @@
     <Compile Include="Coin\Transactions\GenerationTransaction.cs" />
     <Compile Include="Coin\Helpers\Amount.cs" />
     <Compile Include="Common\Repository\Bootstrapper.cs" />
-    <Compile Include="Common\Config\IConfig.cs" />
+    <Compile Include="Common\Configuration\IConfig.cs" />
     <Compile Include="Common\Repository\Registries\ClassRegistry.cs" />
     <Compile Include="Common\Repository\Registries\IRegistry.cs" />
     <Compile Include="Common\Repository\Registries\FactoryRegistry.cs" />
@@ -133,27 +133,27 @@
     <Compile Include="Common\Repository\Registries\Registry.cs" />
     <Compile Include="Crypto\Utils.cs" />
     <Compile Include="Miner\Events\MinerAuthenticationEventArgs.cs" />
-    <Compile Include="Jobs\ExtraNonce.cs" />
-    <Compile Include="Jobs\IExtraNonce.cs" />
-    <Compile Include="Jobs\IJobManagerFactory.cs" />
-    <Compile Include="Jobs\JobCounter.cs" />
-    <Compile Include="Jobs\JobManagerFactory.cs" />
+    <Compile Include="Mining\Jobs\ExtraNonce.cs" />
+    <Compile Include="Mining\Jobs\IExtraNonce.cs" />
+    <Compile Include="Mining\Jobs\IJobManagerFactory.cs" />
+    <Compile Include="Mining\Jobs\JobCounter.cs" />
+    <Compile Include="Mining\Jobs\JobManagerFactory.cs" />
     <Compile Include="Miner\IMinerManager.cs" />
     <Compile Include="Crypto\MerkleTree.cs" />
     <Compile Include="Coin\Daemon\Responses\BlockTemplateTransaction.cs" />
-    <Compile Include="Jobs\IJobManager.cs" />
-    <Compile Include="Jobs\JobManager.cs" />
+    <Compile Include="Mining\Jobs\IJobManager.cs" />
+    <Compile Include="Mining\Jobs\JobManager.cs" />
     <Compile Include="Miner\MinerManager.cs" />
-    <Compile Include="Pool\IPool.cs" />
-    <Compile Include="Pool\Config\IPoolConfig.cs" />
-    <Compile Include="Pool\IPoolFactory.cs" />
-    <Compile Include="Pool\IPoolManager.cs" />
-    <Compile Include="Pool\Pool.cs" />
-    <Compile Include="Pool\PoolFactory.cs" />
-    <Compile Include="Share\IShareManagerFactory.cs" />
-    <Compile Include="Share\IShareManager.cs" />
-    <Compile Include="Share\ShareManagerFactory.cs" />
-    <Compile Include="Share\ShareManager.cs" />
+    <Compile Include="Mining\Pool\IPool.cs" />
+    <Compile Include="Mining\Pool\Config\IPoolConfig.cs" />
+    <Compile Include="Mining\Pool\IPoolFactory.cs" />
+    <Compile Include="Mining\Pool\IPoolManager.cs" />
+    <Compile Include="Mining\Pool\Pool.cs" />
+    <Compile Include="Mining\Pool\PoolFactory.cs" />
+    <Compile Include="Mining\Share\IShareManagerFactory.cs" />
+    <Compile Include="Mining\Share\IShareManager.cs" />
+    <Compile Include="Mining\Share\ShareManagerFactory.cs" />
+    <Compile Include="Mining\Share\ShareManager.cs" />
     <Compile Include="Rpc\Service\IRpcService.cs" />
     <Compile Include="Server\IMiningServer.cs" />
     <Compile Include="Server\Config\IServerConfig.cs" />
@@ -165,8 +165,8 @@
     <Compile Include="Server\Stratum\Notifications\Difficulty.cs" />
     <Compile Include="Server\Vanilla\Config\IVanillaServerConfig.cs" />
     <Compile Include="Server\Vanilla\VanillaMiner.cs" />
-    <Compile Include="Pool\Config\PoolConfig.cs" />
-    <Compile Include="Pool\PoolManager.cs" />
+    <Compile Include="Mining\Pool\Config\PoolConfig.cs" />
+    <Compile Include="Mining\Pool\PoolManager.cs" />
     <Compile Include="Rpc\Service\Http\HttpServiceContext.cs" />
     <Compile Include="Rpc\Service\Http\HttpServiceRequest.cs" />
     <Compile Include="Server\Vanilla\VanillaServer.cs" />
